+++ backend/src/main/java/com/icehufs/icebreaker/domain/auth/domain/entity/UserBan.java
@@ -1,6 +1,10 @@
-package com.icehufs.icebreaker.entity;
+package com.icehufs.icebreaker.domain.auth.domain.entity;
 
 import java.time.LocalDateTime;
+
+import com.icehufs.icebreaker.domain.auth.domain.type.BanDurationEnum;
+import com.icehufs.icebreaker.domain.auth.domain.type.BanReasonEnum;
+import com.icehufs.icebreaker.entity.User;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
