+++ backend/src/main/java/com/icehufs/icebreaker/domain/auth/repostiory/AuthorityRepository.java
@@ -1,0 +1,16 @@
+package com.icehufs.icebreaker.domain.auth.repostiory;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.icehufs.icebreaker.domain.auth.domain.entity.Authority;
+import java.util.List;
+
+
+@Repository
+public interface AuthorityRepository extends JpaRepository<Authority, String> {
+        Authority findByEmail(String email);
+        List<Authority> findByRoleAdminC1(String roleAdminC1);
+        List<Authority> findByRoleAdminC2(String roleAdminC2);
+
+}
