+++ backend/src/main/java/com/icehufs/icebreaker/domain/auth/repostiory/CertificationRepository.java
@@ -1,0 +1,13 @@
+package com.icehufs.icebreaker.domain.auth.repostiory;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.icehufs.icebreaker.domain.auth.domain.entity.Certification;
+
+@Repository
+public interface CertificationRepository extends JpaRepository<Certification, String> {
+
+    Certification findByUserEmail(String userEmail);
+    
+}
